--- a/src/Nethermind/Nethermind.AccountAbstraction/Executor/AABlockProducerTransactionsExecutor.cs
+++ b/src/Nethermind/Nethermind.AccountAbstraction/Executor/AABlockProducerTransactionsExecutor.cs
@@ -48,6 +48,8 @@ namespace Nethermind.AccountAbstraction.Executor
         {
             IEnumerable<Transaction> transactions = GetTransactions(block);
 
+            Evm.Metrics.ResetBlockStats();
+
             int i = 0;
             LinkedHashSet<Transaction> transactionsInBlock = new(ByHashTxComparer.Instance);
             foreach (Transaction transaction in transactions)