--- a/src/Nethermind/Nethermind.Consensus/Processing/BlockProcessor.BlockProductionTransactionsExecutor.cs
+++ b/src/Nethermind/Nethermind.Consensus/Processing/BlockProcessor.BlockProductionTransactionsExecutor.cs
@@ -65,6 +65,8 @@ namespace Nethermind.Consensus.Processing
             {
                 IEnumerable<Transaction> transactions = GetTransactions(block);
 
+                Evm.Metrics.ResetBlockStats();
+
                 int i = 0;
                 LinkedHashSet<Transaction> transactionsInBlock = new(ByHashTxComparer.Instance);
                 foreach (Transaction currentTx in transactions)