--- a/src/Nethermind/Nethermind.Consensus/Processing/ProcessingStats.cs
+++ b/src/Nethermind/Nethermind.Consensus/Processing/ProcessingStats.cs
@@ -115,15 +115,16 @@ namespace Nethermind.Consensus.Processing
                     decimal runMs = (runMicroseconds == 0 ? -1 : runMicroseconds / 1000);
                     if (chunkBlocks > 1)
                     {
-                        _logger.Info($"Processed {block.Number - chunkBlocks + 1,9}...{block.Number,9} | {chunkMs,9:N2} ms  | slot {runMs,7:N0} ms     | recv  {recoveryQueueSize,7:N0} | proc   {blockQueueSize,6:N0}");
+                        _logger.Info($"Processed   {block.Number - chunkBlocks + 1,9}...{block.Number,9} | {chunkMs,9:N2} ms  | slot {runMs,7:N0} ms     | recv  {recoveryQueueSize,7:N0} | proc   {blockQueueSize,6:N0}");
                     }
                     else
                     {
-                        _logger.Info($"Processed   {block.Number,9}           | {chunkMs,9:N2} ms  | slot {runMs,7:N0} ms     | recv  {recoveryQueueSize,7:N0} | proc   {blockQueueSize,6:N0}");
+                        _logger.Info($"Processed     {block.Number,9}           | {chunkMs,9:N2} ms  | slot {runMs,7:N0} ms     | recv  {recoveryQueueSize,7:N0} | proc   {blockQueueSize,6:N0}");
                     }
-                    _logger.Info($"- Block{(chunkBlocks > 1 ? "s" : " ")}         {chunkMGas,7:F2} MGas   | {chunkTx,6:N0}    txs | calls {chunkCalls,6:N0} ({chunkEmptyCalls,3:N0})  | sload {chunkSload,7:N0} | sstore {chunkSstore,6:N0} | creates {chunkCreates,3:N0} ({-(currentSelfDestructs - _lastSelfDestructs),3:N0})");
-                    _logger.Info($"- Throughput     {mgasPerSecond,7:F2} MGas/s | {txps,9:F2} t/s |         {bps,7:F2} b/s");
-                    _logger.Info($"- Ave Throughput {totalMgasPerSecond,7:F2} MGas/s | {totalTxPerSecond,9:F2} t/s |         {totalBlocksPerSecond,7:F2} b/s");
+                    _logger.Info($"- Block{(chunkBlocks > 1 ? "s" : " ")}           {chunkMGas,7:F2} MGas   | {chunkTx,6:N0}    txs | calls {chunkCalls,6:N0} ({chunkEmptyCalls,3:N0})  | sload {chunkSload,7:N0} | sstore {chunkSstore,6:N0} | creates {chunkCreates,3:N0} ({-(currentSelfDestructs - _lastSelfDestructs),3:N0})");
+                    string blockGas = $" Gas gwei: {Evm.Metrics.BlockMinGasPrice:N2} .. {Math.Max(Evm.Metrics.BlockMinGasPrice, Evm.Metrics.BlockEstMedianGasPrice):N2} ({Evm.Metrics.BlockAveGasPrice:N2}) .. {Evm.Metrics.BlockMaxGasPrice:N2}";
+                    _logger.Info($"- Block Throughput {mgasPerSecond,7:F2} MGas/s | {txps,9:F2} t/s |         {bps,7:F2} b/s |{(Evm.Metrics.BlockMinGasPrice != decimal.MaxValue ? blockGas : "")}");
+                    _logger.Info($"- Total Throughput {totalMgasPerSecond,7:F2} MGas/s | {totalTxPerSecond,9:F2} t/s |         {totalBlocksPerSecond,7:F2} b/s | Gas gwei: {Evm.Metrics.MinGasPrice:N2} .. {Math.Max(Evm.Metrics.MinGasPrice, Evm.Metrics.EstMedianGasPrice):N2} ({Evm.Metrics.AveGasPrice:N2}) .. {Evm.Metrics.MaxGasPrice:N2}");
                 }
                 if (_logger.IsTrace)
                 {