… src/Nethermind/Nethermind.Db.Rocks/DbOnTheRocks.cs
… src/Nethermind/Nethermind.Evm/Metrics.cs
… src/Nethermind/Nethermind.Evm/TransactionProcessing/TransactionProcessor.cs
… src/Nethermind/Nethermind.Evm/VirtualMachine.cs
… src/Nethermind/Nethermind.Init/MemoryHintMan.cs
… src/Nethermind/Nethermind.Init/Steps/Migrations/BloomMigration.cs
… src/Nethermind/Nethermind.Init/Steps/Migrations/ReceiptMigration.cs
… src/Nethermind/Nethermind.Merge.Plugin.Test/EngineModuleTests.V1.cs
… src/Nethermind/Nethermind.Merge.Plugin/Data/ForkchoiceStateV1.cs
… src/Nethermind/Nethermind.Merge.Plugin/Handlers/ForkchoiceUpdatedHandler.cs
… src/Nethermind/Nethermind.Merge.Plugin/Handlers/NewPayloadHandler.cs
… src/Nethermind/Nethermind.Merge.Plugin/Synchronization/PeerRefresher.cs
… src/Nethermind/Nethermind.Merge.Plugin/Synchronization/PivotUpdator.cs
… src/Nethermind/Nethermind.Mev/MevBlockProductionTransactionsExecutor.cs
… src/Nethermind/Nethermind.Runner/NLog.config
… src/Nethermind/Nethermind.Synchronization.Test/BlockDownloaderTests.cs
… src/Nethermind/Nethermind.Synchronization.Test/FastSync/StateSyncFeedTestsBase.cs
… src/Nethermind/Nethermind.Synchronization.Test/LatencySyncPeerMock.cs
… src/Nethermind/Nethermind.Synchronization.Test/SyncPeerMock.cs
… src/Nethermind/Nethermind.Synchronization.Test/SyncPeerPoolTests.cs
… src/Nethermind/Nethermind.Synchronization.Test/SyncPeersReportTests.cs
… src/Nethermind/Nethermind.Synchronization.Test/SyncReportTest.cs
… src/Nethermind/Nethermind.Synchronization.Test/SynchronizerTests.cs
… src/Nethermind/Nethermind.Synchronization/FastSync/DetailedProgress.cs
… src/Nethermind/Nethermind.Synchronization/ParallelSync/MultiSyncModeSelector.cs
… src/Nethermind/Nethermind.Synchronization/Peers/SyncPeersReport.cs
… src/Nethermind/Nethermind.Synchronization/Reporting/SyncReport.cs
… src/Nethermind/Nethermind.Synchronization/SnapSync/Pivot.cs
… src/Nethermind/Nethermind.Synchronization/SnapSync/ProgressTracker.cs
… src/Nethermind/Nethermind.Trie/Pruning/MemoryLimit.cs