Contributes to https://github.com/NethermindEth/nethermind/issues/5758
Resolves https://github.com/NethermindEth/nethermind/issues/268

Improve logging info for blocks #5769 with the additions of min, median, mean and max gas in gwei

## Changes

- Is currently hard to read a block log or derive specific block information from it (how many txns, how much gas etc); also the mix in of total is confusing. If a block takes a long time its not providing any additional detail on what is happening in the block:

Block validation  (Left current, right new)

![image](https://github.com/NethermindEth/nethermind/assets/1142958/473d0a7c-6b3b-4b7d-a4d2-749db52d5600)

Syncing (Left current, right new)

![image](https://github.com/NethermindEth/nethermind/assets/1142958/832f294d-9e34-4784-8254-322507e17241)

<img width="812" alt="image" src="https://github.com/NethermindEth/nethermind/assets/1142958/1bf15c10-cd26-42d6-8854-868244fbeeb3">

```
Processed   17320937 |      231ms of  12,376ms, mgasps  104.59 total   96.82, tps  981.08 total  973.89, bps    4.32 total    6.38, recv queue 0, proc queue 0 
```
This changes it to the following for single blocks (with surrounding other logs) and aligns the numeric columns so easier to read

7 blocks

Before (49 lines)
```
Received a new payload: 17453905 (0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3)
Processed   17453905 |      195ms of  11,770ms, mgasps   60.54 total   51.42, tps  810.04 total  511.55, bps    5.13 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453905 (0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3).
Received: ForkchoiceState: (HeadBlockHash: 0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453905 (0x0a8eba...79f1b3) was processed.
Block 0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453906 (0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa)
Processed   17453906 |      578ms of  12,396ms, mgasps   51.91 total   51.42, tps  266.51 total  510.38, bps    1.73 total    3.34, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453906 (0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa).
Received: ForkchoiceState: (HeadBlockHash: 0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453906 (0xa24f23...9cf4fa) was processed.
Block 0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453907 (0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7)
Processed   17453907 |      270ms of  12,732ms, mgasps   48.14 total   51.42, tps  674.55 total  510.74, bps    3.71 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453907 (0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7).
Received: ForkchoiceState: (HeadBlockHash: 0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453907 (0x8b01ba...0f28e7) was processed.
Block 0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453908 (0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a)
Processed   17453908 |      253ms of  11,043ms, mgasps   45.96 total   51.41, tps  736.55 total  511.21, bps    3.96 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453908 (0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a).
Received: ForkchoiceState: (HeadBlockHash: 0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453908 (0xb48a98...58556a) was processed.
Block 0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453909 (0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274)
Processed   17453909 |      267ms of  13,569ms, mgasps   36.53 total   51.37, tps  483.29 total  511.15, bps    3.75 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453909 (0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274).
Received: ForkchoiceState: (HeadBlockHash: 0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453909 (0x064c26...a6e274) was processed.
Block 0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453910 (0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383)
Processed   17453910 |      839ms of  11,321ms, mgasps   35.66 total   51.27, tps  180.02 total  508.88, bps    1.19 total    3.33, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453910 (0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383).
Received: ForkchoiceState: (HeadBlockHash: 0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453910 (0x9beea1...361383) was processed.
Block 0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453911 (0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e)
Processed   17453911 |      336ms of  10,863ms, mgasps   23.49 total   51.19, tps  392.82 total  508.56, bps    2.98 total    3.33, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453911 (0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e).
Received: ForkchoiceState: (HeadBlockHash: 0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453911 (0x36eba8...e7562e) was processed.
Block 0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
```
After 42 lines (and they are interesting and readable lines)
```
Received a new payload: 17454194 (0xe8b6d7865fabd44e342d86069499cd34b3490013ba6383937a244ccb3082de25)
Processed    17454193... 17454194 |    655.39 ms  | slot   3,963 ms     | Gas gwei: 16.28 .. 16.28 (20.49) .. 32.90
- Blocks             37.60 MGas   |    319    txs | calls  1,270 ( 48)  | sload   3,742 | sstore  1,894 | create 100
- Block Throughput   57.38 MGas/s |    486.73 t/s |            3.05 b/s | recv        0 | proc        0
Received: ForkChoice: (Head: 0xe8b6d7865fabd44e342d86069499cd34b3490013ba6383937a244ccb3082de25, Safe: 0xf31207...32404a, Finalized: 0x00c8fb...61b2d0)
Block 0xe8b6d7865fabd44e342d86069499cd34b3490013ba6383937a244ccb3082de25 was set as head.
Received a new payload: 17454195 (0x4e0fa605d30b8f3334b2afcda1e6bb17f42f5050166c299f01d316e103053abd)
Processed      17454195           |    500.31 ms  | slot   7,627 ms     | Gas gwei: 15.50 .. 15.52 (20.77) .. 32.90
- Block              30.00 MGas   |    143    txs | calls    813 (  4)  | sload   1,846 | sstore  1,369 | create 129
- Block Throughput   59.96 MGas/s |    285.83 t/s |            2.00 b/s | recv        0 | proc        0
Received: ForkChoice: (Head: 0x4e0fa605d30b8f3334b2afcda1e6bb17f42f5050166c299f01d316e103053abd, Safe: 0xf31207...32404a, Finalized: 0x00c8fb...61b2d0)
Block 0x4e0fa605d30b8f3334b2afcda1e6bb17f42f5050166c299f01d316e103053abd was set as head.
Received a new payload: 17454196 (0x7c4a24b0997bece6aa432867983b5d249d1b07a88f9874448049dd1bc1668f11)
Processed      17454196           |    166.19 ms  | slot   9,798 ms     | Gas gwei: 17.44 .. 17.44 (20.25) .. 35.43
- Block               8.42 MGas   |    151    txs | calls    220 (  5)  | sload   1,182 | sstore    430 | create   0
- Block Throughput   50.67 MGas/s |    908.62 t/s |            6.02 b/s | recv        0 | proc        0
Received: ForkChoice: (Head: 0x7c4a24b0997bece6aa432867983b5d249d1b07a88f9874448049dd1bc1668f11, Safe: 0xf31207...32404a, Finalized: 0x00c8fb...61b2d0)
Block 0x7c4a24b0997bece6aa432867983b5d249d1b07a88f9874448049dd1bc1668f11 was set as head.
Received a new payload: 17454197 (0xa66ab2d48e36fe8190a2a165551dfc2904f8ab1e9d124fd079219051e241f550)
Processed      17454197           |    185.18 ms  | slot  11,638 ms     | Gas gwei: 16.48 .. 16.51 (20.97) .. 53.00
- Block               8.73 MGas   |    145    txs | calls    335 ( 12)  | sload   1,587 | sstore    485 | create   0
- Block Throughput   47.12 MGas/s |    783.02 t/s |            5.40 b/s | recv        0 | proc        0
Received: ForkChoice: (Head: 0xa66ab2d48e36fe8190a2a165551dfc2904f8ab1e9d124fd079219051e241f550, Safe: 0xf31207...32404a, Finalized: 0x00c8fb...61b2d0)
Block 0xa66ab2d48e36fe8190a2a165551dfc2904f8ab1e9d124fd079219051e241f550 was set as head.
Received a new payload: 17454198 (0x04009d9c0d49d6b3e7720844797d2cbf3d979af89c192a06953101bb791fb3f5)
Processed      17454198           |    139.43 ms  | slot  12,819 ms     | Gas gwei: 15.62 .. 15.71 (20.02) .. 87.88
- Block               9.88 MGas   |    145    txs | calls    343 ( 18)  | sload   1,500 | sstore    430 | create   1
- Block Throughput   70.83 MGas/s |   1039.97 t/s |            7.17 b/s | recv        0 | proc        0
Received: ForkChoice: (Head: 0x04009d9c0d49d6b3e7720844797d2cbf3d979af89c192a06953101bb791fb3f5, Safe: 0xf31207...32404a, Finalized: 0x00c8fb...61b2d0)
Block 0x04009d9c0d49d6b3e7720844797d2cbf3d979af89c192a06953101bb791fb3f5 was set as head.
Received a new payload: 17454199 (0xf0265db12918a19f028bb5281305ceeb2ca36f33892804dfa1416090b375eb03)
Processed      17454199           |    724.83 ms  | slot  12,106 ms     | Gas gwei: 14.95 .. 15.09 (19.75) .. 61.08
- Block              26.18 MGas   |    158    txs | calls  1,656 ( 21)  | sload   4,084 | sstore  2,826 | create  50
- Block Throughput   36.11 MGas/s |    217.98 t/s |            1.38 b/s | recv        0 | proc        0
Received: ForkChoice: (Head: 0xf0265db12918a19f028bb5281305ceeb2ca36f33892804dfa1416090b375eb03, Safe: 0xf31207...32404a, Finalized: 0x00c8fb...61b2d0)
Block 0xf0265db12918a19f028bb5281305ceeb2ca36f33892804dfa1416090b375eb03 was set as head.
Received a new payload: 17454200 (0x3ead96ad6e116d0ad82222cfcd52c829fee7058a9b08d9ce4471b116bb58b44d)
Processed      17454200           |    192.23 ms  | slot  12,999 ms     | Gas gwei: 16.34 .. 23.18 (24.32) .. 655.00
- Block               9.70 MGas   |    175    txs | calls    334 ( 10)  | sload   1,393 | sstore    483 | create   0
- Block Throughput   50.46 MGas/s |    910.37 t/s |            5.20 b/s | recv        0 | proc        0
Received: ForkChoice: (Head: 0x3ead96ad6e116d0ad82222cfcd52c829fee7058a9b08d9ce4471b116bb58b44d, Safe: 0xf31207...32404a, Finalized: 0x00c8fb...61b2d0)
Block 0x3ead96ad6e116d0ad82222cfcd52c829fee7058a9b08d9ce4471b116bb58b44d was set as head.
```

This is much easier to read and allows the logs to assist in investigations on block processing issues

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [ ] Optimization
- [x] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [x] Other: Logs

## Testing

#### Requires testing

- [ ] Yes
- [x] No

#### If yes, did you write tests?

- [ ] Yes
- [x] No