@benaadams 2023-06-11
7 blocks

Before (49 lines)
```
Received a new payload: 17453905 (0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3)
Processed   17453905 |      195ms of  11,770ms, mgasps   60.54 total   51.42, tps  810.04 total  511.55, bps    5.13 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453905 (0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3).
Received: ForkchoiceState: (HeadBlockHash: 0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453905 (0x0a8eba...79f1b3) was processed.
Block 0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x0a8ebacd435da5f713651c04d6297f56f80b0bbc511c9d4bc3acf39df579f1b3, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453906 (0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa)
Processed   17453906 |      578ms of  12,396ms, mgasps   51.91 total   51.42, tps  266.51 total  510.38, bps    1.73 total    3.34, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453906 (0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa).
Received: ForkchoiceState: (HeadBlockHash: 0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453906 (0xa24f23...9cf4fa) was processed.
Block 0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0xa24f23df39f9f73daf2d618dca32df6d655dbbdf7a08bd81c62894dcf59cf4fa, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453907 (0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7)
Processed   17453907 |      270ms of  12,732ms, mgasps   48.14 total   51.42, tps  674.55 total  510.74, bps    3.71 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453907 (0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7).
Received: ForkchoiceState: (HeadBlockHash: 0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453907 (0x8b01ba...0f28e7) was processed.
Block 0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x8b01bad4264fe9aafe28843c3b0f12f46e9aa6252c512d267a9c675e8f0f28e7, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453908 (0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a)
Processed   17453908 |      253ms of  11,043ms, mgasps   45.96 total   51.41, tps  736.55 total  511.21, bps    3.96 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453908 (0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a).
Received: ForkchoiceState: (HeadBlockHash: 0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453908 (0xb48a98...58556a) was processed.
Block 0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0xb48a98ec55695b740cdfd42b4c83b3fad9425c62f263a3aea278d6630e58556a, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453909 (0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274)
Processed   17453909 |      267ms of  13,569ms, mgasps   36.53 total   51.37, tps  483.29 total  511.15, bps    3.75 total    3.35, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453909 (0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274).
Received: ForkchoiceState: (HeadBlockHash: 0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453909 (0x064c26...a6e274) was processed.
Block 0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x064c2665e5f660bb1814caa5f3bee562174cdf5288952e7561531e1760a6e274, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453910 (0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383)
Processed   17453910 |      839ms of  11,321ms, mgasps   35.66 total   51.27, tps  180.02 total  508.88, bps    1.19 total    3.33, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453910 (0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383).
Received: ForkchoiceState: (HeadBlockHash: 0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453910 (0x9beea1...361383) was processed.
Block 0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383 was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x9beea1dd9af771c0fd8c6f6c9317e7b5614ae1f500789b288abe104cf6361383, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
Received a new payload: 17453911 (0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e)
Processed   17453911 |      336ms of  10,863ms, mgasps   23.49 total   51.19, tps  392.82 total  508.56, bps    2.98 total    3.33, recv queue 0, proc queue 0
Valid. Result of a new payload: 17453911 (0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e).
Received: ForkchoiceState: (HeadBlockHash: 0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba)
FCU - block 17453911 (0x36eba8...e7562e) was processed.
Block 0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e was set as head.
Valid. Request: ForkchoiceState: (HeadBlockHash: 0x36eba8900400d7485c02ee0f03282f345aae909e3236e057506805ae6ee7562e, SafeBlockHash: 0x852a71291cade5a723f8a3d78c1b9159de728735ae5e79f5be3271d9a5f7798e, FinalizedBlockHash: 0x17a3d74c1bac6357f18278cb6e73abfc3e7c5cb548ed01878cee9d9932e342ba) .
```
After 42 lines
```
Received a new payload: 17454110 (0x672e73d816caf9aeb229411a0824fb7ed768ff11fe18051cc052814cf1b3f981)
Processed      17454110           |    212.78 ms  | slot  13,855 ms     | recv        0 | proc        0
- Block              16.91 MGas   |    173    txs | calls    607 ( 14)  | sload   2,561 | sstore    852 | create   0
- Block Throughput   79.49 MGas/s |    813.05 t/s |            4.70 b/s | Gas gwei: 15.68 .. 15.86 (20.71) .. 129.38
Received: ForkChoice: (Head: 0x672e73d816caf9aeb229411a0824fb7ed768ff11fe18051cc052814cf1b3f981, Safe: 0xd4a93a...ba1902, Finalized: 0x641962...b1ccae)
Block 0x672e73d816caf9aeb229411a0824fb7ed768ff11fe18051cc052814cf1b3f981 was set as head.
Received a new payload: 17454111 (0xa8c9ab0dbcaebdde702e944e299782802db08e4a9d2c11fc6461e8597f2662fc)
Processed      17454111           |    512.84 ms  | slot  10,072 ms     | recv        0 | proc        0
- Block              29.04 MGas   |    135    txs | calls    851 (  7)  | sload   2,072 | sstore  1,372 | create 128
- Block Throughput   56.62 MGas/s |    263.24 t/s |            1.95 b/s | Gas gwei: 15.93 .. 20.95 (28.43) .. 1,144.55
Received: ForkChoice: (Head: 0xa8c9ab0dbcaebdde702e944e299782802db08e4a9d2c11fc6461e8597f2662fc, Safe: 0xd4a93a...ba1902, Finalized: 0x641962...b1ccae)
Block 0xa8c9ab0dbcaebdde702e944e299782802db08e4a9d2c11fc6461e8597f2662fc was set as head.
Received a new payload: 17454112 (0x149584cd8c43bc72bd3b6fe7cf6e2e09e5e9efae17acc34844d0c9651e1a4f1c)
Processed      17454112           |    352.24 ms  | slot  11,740 ms     | recv        0 | proc        0
- Block              12.35 MGas   |    158    txs | calls    498 ( 38)  | sload   2,072 | sstore    633 | create   0
- Block Throughput   35.05 MGas/s |    448.56 t/s |            2.84 b/s | Gas gwei: 17.79 .. 21.79 (21.28) .. 47.79
Received: ForkChoice: (Head: 0x149584cd8c43bc72bd3b6fe7cf6e2e09e5e9efae17acc34844d0c9651e1a4f1c, Safe: 0xd4a93a...ba1902, Finalized: 0x641962...b1ccae)
Block 0x149584cd8c43bc72bd3b6fe7cf6e2e09e5e9efae17acc34844d0c9651e1a4f1c was set as head.
Received a new payload: 17454113 (0x8bef49bea2923d2d5def9f5f8a5821219e2799e68c2286dba0835e1760adb42d)
Processed      17454113           |    193.65 ms  | slot  12,612 ms     | recv        0 | proc        0
- Block              11.06 MGas   |    178    txs | calls    388 ( 29)  | sload   1,577 | sstore    543 | create   0
- Block Throughput   57.10 MGas/s |    919.20 t/s |            5.16 b/s | Gas gwei: 17.40 .. 17.40 (21.04) .. 66.00
Received: ForkChoice: (Head: 0x8bef49bea2923d2d5def9f5f8a5821219e2799e68c2286dba0835e1760adb42d, Safe: 0xd4a93a...ba1902, Finalized: 0x641962...b1ccae)
Block 0x8bef49bea2923d2d5def9f5f8a5821219e2799e68c2286dba0835e1760adb42d was set as head.
Received a new payload: 17454114 (0xa81f2d1bd0929362179a6bfc72cfb1e4f4a0be1ddc5e5fabb51929b3ea9a4ca0)
Processed      17454114           |    207.96 ms  | slot  12,185 ms     | recv        0 | proc        0
- Block               8.46 MGas   |    124    txs | calls    362 ( 13)  | sload   1,428 | sstore    595 | create   0(-10)
- Block Throughput   40.67 MGas/s |    596.28 t/s |            4.81 b/s | Gas gwei: 16.83 .. 16.83 (19.72) .. 42.22
Received: ForkChoice: (Head: 0xa81f2d1bd0929362179a6bfc72cfb1e4f4a0be1ddc5e5fabb51929b3ea9a4ca0, Safe: 0xd4a93a...ba1902, Finalized: 0x641962...b1ccae)
Block 0xa81f2d1bd0929362179a6bfc72cfb1e4f4a0be1ddc5e5fabb51929b3ea9a4ca0 was set as head.
Received a new payload: 17454115 (0x00c8fb49ac09972e89cb7f1bc68c1b14bcf4ae00fd6277dbe30aaa4d1e61b2d0)
Processed      17454115           |    166.30 ms  | slot  12,980 ms     | recv        0 | proc        0
- Block              19.26 MGas   |    144    txs | calls    394 ( 29)  | sload   1,396 | sstore    513 | create   1
- Block Throughput  115.82 MGas/s |    865.89 t/s |            6.01 b/s | Gas gwei: 15.91 .. 15.91 (20.61) .. 150.57
Received: ForkChoice: (Head: 0x00c8fb49ac09972e89cb7f1bc68c1b14bcf4ae00fd6277dbe30aaa4d1e61b2d0, Safe: 0x157eed...b40bc4, Finalized: 0xd4a93a...ba1902)
Block 0x00c8fb49ac09972e89cb7f1bc68c1b14bcf4ae00fd6277dbe30aaa4d1e61b2d0 was set as head.
Received a new payload: 17454116 (0xebfd1c8f4056e9bb5bd0b3900d4bb3a3ff1c5836de46d787891b0adeda6c085b)
Processed      17454116           |    198.86 ms  | slot  10,702 ms     | recv        0 | proc        0
- Block               8.92 MGas   |    158    txs | calls    177 (  9)  | sload   1,143 | sstore    416 | create   1
- Block Throughput   44.88 MGas/s |    794.54 t/s |            5.03 b/s | Gas gwei: 16.47 .. 16.47 (20.04) .. 66.47
Received: ForkChoice: (Head: 0xebfd1c8f4056e9bb5bd0b3900d4bb3a3ff1c5836de46d787891b0adeda6c085b, Safe: 0x157eed...b40bc4, Finalized: 0xd4a93a...ba1902)
Block 0xebfd1c8f4056e9bb5bd0b3900d4bb3a3ff1c5836de46d787891b0adeda6c085b was set as head.
```